// providers/socket-types.tsx
export interface RecipientResponseProps {
	profile: Profile;
	socket: SocketProps;
	_id: string;
	firstName: string;
	lastName: string;
	score: number;
	profileImage: ProfileImage;
	type?: Roles;
	role?: Roles;
}
// ...
export type CombinedMessageProps = Partial<MessageResponseProps> & Partial<TemporaryBubbleProps>;

export interface ConversationResponseProps {
	_id: string;
	type: string;
	recipients: RecipientResponseProps[];
	messages: CombinedMessageProps[];
	createdAt: string;
	updatedAt: string;
	__v: number;
	// conversation.type !== "DIRECT" the below properties will be available
	title?: string;
	description?: string;
}

export type CombinedConversationResponseProps = Partial<ConversationResponseProps> & Partial<PopUpChatDataProps>;

export interface GetAllChatsResponseProps {
	messages: CombinedConversationResponseProps[];
}
export interface GetAllChatsResponseProps2 {
	chats: {
		messages: ConversationResponseProps[];
	};
	_id: string;
	createdAt: string;
	updatedAt: string;
	recipients: RecipientResponseProps[];
}

// ====== ALL CHATS ===== //

// ====== POP-UP MESSAGE ===== //

interface PopUpMessageProps {
	attachments: string[];
	content: string;
	conversation: string;
	createdAt: string;
	_id: string;
	user: string;
	type: string;
	seen: string | null;
	readBy: string[];
	updatedAt: string;
	__v: number;
}

interface Bio {
	title: string;
	description: string;
}

interface Profile {
	bio: Bio;
}

interface PopUpProfileImageProps {
	_id: string;
	url: string;
}

export interface PopUpRecipientProps {
	profile: Profile;
	socket: SocketProps;
	_id: string;
	firstName: string;
	lastName: string;
	score: number;
	profileImage: PopUpProfileImageProps;
}

export interface PopUpChatDataProps {
	messages: PopUpMessageProps[];
	totalMessagesCount: number;
}